Refactoring:
- Moved seven_segment to a mod inside the clock state
- Moved some static buffers out of static, and put inside their
respsective structures
- added syscall tests, tests for more integral systems still need to be
written
- when the itm feature is enabled, use panic_itm instead
