+++ Cargo.toml
@@ -44,7 +44,7 @@ version = "0.4.1"
 [dependencies.stm32l4xx-hal]
 version = "0.3.6"
 # path = "../../hal/stm32l432xx-hal"
-# git = "https://github.com/mabezdev/stm32l4xx-hal"
+git = "https://github.com/mabezdev/stm32l4xx-hal"
 features = ["rt", "stm32l4x2"]
 
 [dependencies.crc]
