+++ src/application/mod.rs
@@ -1,5 +1,4 @@
 pub mod application_manager;
 pub mod display_manager;
 pub mod states;
-pub mod seven_segment;
