+++ src/lib.rs
@@ -2,7 +2,11 @@
 //!
 //! Exposes all the types the sdk may want to use, which the kernel provides
 
-#![no_std]
+#![cfg_attr(not(test), no_std)]
+
+#[cfg(test)]
+#[macro_use]
+extern crate std;
 
 #[macro_use]
 extern crate cortex_m;
